backend/BP/requirements.txt
asgiref==3.3.1
Django==3.1.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
